Address bayardneville's 2026-05-12 PR feedback round
Six items from the second review pass:

F1 (setup-template.md) — Template 2 partial-config crash. Replaced
os.environ[...] reads of MCD_DEFAULT_API_ID/TOKEN with os.environ.get +
a logging.warning skip-and-log block, mirroring the existing endpoint
presence check. A Lambda with OTEL_ENDPOINT set but credentials missing
now runs uninstrumented instead of crashing at cold start with KeyError.

F2 (setup-template.md) — concrete serverless splice for the
prompts-disabled variant. Added a CRITICAL callout plus a full patched
Template 2 snippet showing the three-line os.environ.setdefault block at
module scope (between `import os` and any
`opentelemetry.instrumentation.*` import). Previous "apply at the top of
the chosen Section 1 template" wording invited customers to drop the
line inside `init_tracing()`, which is a no-op because the instrumentors
read the env var at import time.

F3 (detect_libraries.py + instrumentor_map.json + references) — biggest
change. Deleted instrumentor_map.json and rewrote detect_libraries.py as
a thin discovery layer. Output is now `dependencies` (sorted raw pip
names) + runtime/serverless_signals/existing_setup. AI-library matching
is the LLM's job, using fetch_sdk_docs.py's live PyPI list as the
ground-truth supported set. Removed AMBIGUOUS_AWS_* hardcoded
disambiguation — ambiguous deps (boto3, google-cloud-aiplatform, etc.)
are now surfaced via user prompt per library-detection.md section 4.
Moved the wrapt<2 transitive-pin guidance to troubleshooting.md as a
symptom-driven fix (TypeError: wrap_function_wrapper() got an unexpected
keyword argument 'module') rather than baking it into every install
diff. Updated library-detection.md, workflow.md, and SKILL.md to reflect
the new contract.

F7 (setup-template.md) — env-var naming consistency. Standardized on
OTEL_ENDPOINT across all four templates (Templates 2/3/4 already used
it; Template 1 + the prompts-disabled variant were updated from
MC_OTEL_ENDPOINT). Also renamed Template 1's local variable from the
uppercase MC_OTEL_ENDPOINT to lowercase otel_endpoint so the env var
name and the local variable name aren't textually identical.

F9 (test reorg) — moved skills/instrument-agent/scripts/test_*.py and
test_fixtures/ to a sibling skills/instrument-agent/tests/ directory
with the fixtures renamed to tests/fixtures/. Updated path constants in
both test files, sys.path.insert in test_fetch_sdk_docs.py, and the CI
workflow paths in .github/workflows/instrument-agent-tests.yml. CI
`paths:` filter (skills/instrument-agent/**) unchanged. Both test suites
still pass after the move (53 + 33 = 86 checks).

redaction.md:27 — fixed factual claim about the default OTLP endpoint.
The SDK has no built-in default; the customer always supplies it. The
templates resolve it from an OTEL_ENDPOINT env var so they can flip
between MC-hosted and self-hosted without code changes.
