--- skills/instrument-agent/scripts/instrumentor_map.json
@@ -1,88 +1,0 @@
-{
-  "$schema_note": "Static detection map for `detect_libraries.py`. Maps installed Python package names (covers_dependencies) to Monte Carlo OpenTelemetry instrumentor packages so the skill can recognize what to install for a given project offline. This file is NOT a fallback for `fetch_sdk_docs.py`: PyPI is the source of truth for the live supported-instrumentor list. The version_constraint here is a last-known-compatible value used to pin transitive resolution; re-run fetch_sdk_docs.py to confirm the current PyPI value. additional_pins are transitive constraints PyPI metadata does not expose (e.g. wrapt<2 for OpenLLMetry instrumentors that pass module= to wrap_function_wrapper, which wrapt 2.x renamed to target=).",
-  "supported_instrumentors": [
-    {
-      "library": "langchain",
-      "package": "opentelemetry-instrumentation-langchain",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "langchain",
-        "langchain-core",
-        "langchain-community",
-        "langchain-openai",
-        "langchain-anthropic",
-        "langgraph"
-      ]
-    },
-    {
-      "library": "langgraph",
-      "package": "opentelemetry-instrumentation-langchain",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "langgraph"
-      ],
-      "notes": "LangGraph traces are produced by the same opentelemetry-instrumentation-langchain package."
-    },
-    {
-      "library": "openai",
-      "package": "opentelemetry-instrumentation-openai",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "openai"
-      ]
-    },
-    {
-      "library": "anthropic",
-      "package": "opentelemetry-instrumentation-anthropic",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "anthropic"
-      ]
-    },
-    {
-      "library": "crewai",
-      "package": "opentelemetry-instrumentation-crewai",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "crewai"
-      ]
-    },
-    {
-      "library": "bedrock",
-      "package": "opentelemetry-instrumentation-bedrock",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "boto3",
-        "botocore",
-        "aioboto3"
-      ],
-      "notes": "boto3 covers many AWS services; detect_libraries.py routes boto3 matches to `unsupported` rather than `detected` since it cannot tell from deps alone whether the customer is calling Bedrock. The skill should ask the user before installing this instrumentor."
-    },
-    {
-      "library": "sagemaker",
-      "package": "opentelemetry-instrumentation-sagemaker",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "sagemaker"
-      ],
-      "notes": "Verify on PyPI at runtime — package availability has historically been less stable than the LangChain/OpenAI/Anthropic family."
-    },
-    {
-      "library": "vertexai",
-      "package": "opentelemetry-instrumentation-vertexai",
-      "version_constraint": "<=0.53.4",
-      "additional_pins": ["wrapt<2"],
-      "covers_dependencies": [
-        "google-cloud-aiplatform"
-      ],
-      "notes": "Verify on PyPI at runtime — package name and version constraint may have changed."
-    }
-  ]
-}
