+++ skills/instrument-agent/tests/fixtures/existing-setup/requirements.txt
