+++ skills/instrument-agent/tests/fixtures/mixed-requirements-pyproject/pyproject.toml
