+++ skills/instrument-agent/tests/fixtures/mixed-requirements-pyproject/requirements.txt
